Testing: Move test_user_sitepackages_in_pathlist to its own file
