+++ spyder_kernels/customize/tests/test_utils.py
@@ -1,0 +1,23 @@
+# -*- coding: utf-8 -*-
+# -----------------------------------------------------------------------------
+# Copyright (c) 2009- Spyder Kernels Contributors
+#
+# Licensed under the terms of the MIT License
+# (see spyder_kernels/__init__.py for details)
+# -----------------------------------------------------------------------------
+
+import sys
+
+from spyder_kernels.customize.utils import create_pathlist
+
+
+def test_user_sitepackages_in_pathlist():
+    """Test that we include user site-packages in pathlist."""
+    if sys.platform.startswith('linux'):
+        user_path = 'local'
+    elif sys.platform == 'darwin':
+        user_path = '/Users/travis/.local'
+    else:
+        user_path = 'Roaming'
+
+    assert any([user_path in path for path in create_pathlist()])
